requirements.txt
django-ninja
django-filter
django-cors-headers
markdown
python-dotenv
aiohttp
requests